Some sign-up development
The plan is to enable users to give conditional access to their
Notice account using Shamir's Secret Sharing.
 This will cover the situation where a user dies, but using SSS
they can give part of the key to each friend and their lawyer;
That group would be able to access the account without having to
convince the company hosting it or the sysadmin that they should
have access.