Aliases now work and Alias details store info
Alias details show how users can store data, (passphrases)
without divulging anything to the storage system.
The passphrase is encrypted using AES in javascript by the
browser _BEFORE_ it is sent to the server. So even over
an unencrypted connection, if the POST is intercepted
then the passphrase is still protected.