Added _get_template function.
This enables any Notice module to fetch a template or fragment
and optionally have it parsed for inclusion. This is important
for at least two reasons. 1. It enables good policy of having
NO html in the Control modules and 2. This works with Notice::Pages
so that admin users, (or anyone that you enable) can edit those
templates without having to edit the underlying code or needing
access to that code.