Example of backwardly compatible ajax!
This module is far from ready, but does contain a useful example.
It shows how a module can maintain old runmodes, (for those that
have them bookmarked) and still move into the 21century to support
ajax. (It also encorages smaller functions so that anything can
be called via ajax.)

The next step is to move this back into the core so that we can
have one standard AJAX function that can not only call can function
but any function within any module, (while maintaining all of the
user access features.)